+++ src/sll/api/error.c
@@ -25,15 +25,6 @@ __API_FUNC(error_get_call_stack){
 	sll_array_create(c_st->l-((sll_array_length_t)a),out);
 	for (sll_call_stack_size_t i=0;i<c_st->l-a;i++){
 		const sll_call_stack_frame_t* fr=c_st->dt+i;
-		sll_object_t* r_dt=sll_create_object(SLL_OBJECT_TYPE_ARRAY);
-		sll_array_create(2,&(r_dt->dt.a));
-		r_dt->dt.a.v[0]=sll_int_to_object(fr->_ii);
-		r_dt->dt.a.v[1]=sll_int_to_object(fr->_s);
-		sll_object_t* fr_o=sll_create_object(SLL_OBJECT_TYPE_ARRAY);
-		sll_array_create(3,&(fr_o->dt.a));
-		fr_o->dt.a.v[0]=sll_string_to_object_pointer(fr->nm,SLL_MAX_STRING_LENGTH);
-		fr_o->dt.a.v[1]=sll_instruction_to_location(fr->_ii);
-		fr_o->dt.a.v[2]=r_dt;
-		out->v[i]=fr_o;
+		out->v[i]=sll_new_object(SLL_CHAR("[SO[ii]]"),fr->nm,sll_instruction_to_location(fr->_ii),fr->_ii,fr->_s);
 	}
 }