+++ src/sll/new_object.c
@@ -1,0 +1,133 @@
+#include <sll/_internal/common.h>
+#include <sll/allocator.h>
+#include <sll/array.h>
+#include <sll/common.h>
+#include <sll/map.h>
+#include <sll/memory.h>
+#include <sll/object.h>
+#include <sll/static_object.h>
+#include <sll/string.h>
+#include <sll/types.h>
+#include <sll/var_arg.h>
+#include <stdarg.h>
+
+
+
+static sll_object_t* _build_single(const sll_char_t** t,sll_string_length_t* tl,sll_var_arg_list_t* va){
+	if (!(*tl)){
+		return SLL_ACQUIRE_STATIC_INT(0);
+	}
+	(*tl)--;
+	sll_char_t st=**t;
+	(*t)++;
+	switch (st){
+		case '1':
+			return SLL_ACQUIRE_STATIC_INT(1);
+		case 'i':
+			return sll_int_to_object(sll_var_arg_get_int(va));
+		case 'f':
+			return sll_float_to_object(sll_var_arg_get_float(va));
+		case 'c':
+			return sll_char_to_object(sll_var_arg_get_char(va));
+		case 's':
+			return sll_string_to_object((const sll_string_t*)sll_var_arg_get(va));
+		case 'S':
+			{
+				const sll_char_t* ptr=sll_var_arg_get(va);
+				return (ptr?sll_string_to_object_pointer(ptr,SLL_MAX_STRING_LENGTH):sll_string_to_object(NULL));
+			}
+		case 'L':
+			{
+				const sll_char_t* ptr=sll_var_arg_get(va);
+				sll_string_length_t len=(sll_string_length_t)sll_var_arg_get_int(va);
+				return (ptr&&len?sll_string_to_object_pointer(ptr,len):sll_string_to_object(NULL));
+			}
+		case '[':
+			{
+				sll_object_t* o=sll_create_object(SLL_OBJECT_TYPE_ARRAY);
+				sll_array_create(0,&(o->dt.a));
+				while (tl&&**t!=']'){
+					o->dt.a.l++;
+					sll_allocator_resize((void**)(&(o->dt.a.v)),o->dt.a.l*sizeof(sll_object_t*));
+					o->dt.a.v[o->dt.a.l-1]=_build_single(t,tl,va);
+				}
+				if (tl){
+					(*tl)--;
+					(*t)++;
+				}
+				return o;
+			}
+		case '<':
+			{
+				sll_object_t* o=sll_create_object(SLL_OBJECT_TYPE_MAP);
+				sll_map_create(0,&(o->dt.m));
+				sll_bool_t val=0;
+				while (tl&&**t!='>'){
+					if (!val){
+						o->dt.m.l++;
+						o->dt.m.v=sll_reallocate(o->dt.m.v,(o->dt.m.l<<1)*sizeof(sll_object_t*));
+					}
+					o->dt.m.v[((o->dt.m.l-1)<<1)+val]=_build_single(t,tl,va);
+					val=!val;
+				}
+				if (val){
+					o->dt.m.v[(o->dt.m.l<<1)-1]=SLL_ACQUIRE_STATIC_INT(0);
+				}
+				if (tl){
+					(*tl)--;
+					(*t)++;
+				}
+				return o;
+			}
+		case 'O':
+		case 'R':
+			{
+				sll_object_t* o=(sll_object_t*)sll_var_arg_get(va);
+				if (!o){
+					return SLL_ACQUIRE_STATIC_INT(0);
+				}
+				if (st=='R'){
+					SLL_ACQUIRE(o);
+				}
+				return o;
+			}
+	}
+	return SLL_ACQUIRE_STATIC_INT(0);
+}
+
+
+
+__SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_object_t* sll_new_object(const sll_char_t* t,...){
+	va_list va;
+	va_start(va,t);
+	sll_var_arg_list_t dt={
+		SLL_VAR_ARG_LIST_TYPE_C,
+		{
+			.c=&va
+		}
+	};
+	sll_object_t* o=sll_new_object_list(t,sll_string_length_unaligned(t),&dt);
+	va_end(va);
+	return o;
+}
+
+
+
+__SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_object_t* sll_new_object_list(const sll_char_t* t,sll_string_length_t tl,sll_var_arg_list_t* va){
+	if (!tl){
+		return SLL_ACQUIRE_STATIC_INT(0);
+	}
+	sll_object_t* e=_build_single(&t,&tl,va);
+	if (!tl){
+		return e;
+	}
+	sll_object_t* o=sll_create_object(SLL_OBJECT_TYPE_ARRAY);
+	sll_array_create(1,&(o->dt.a));
+	o->dt.a.v[0]=e;
+	do{
+		o->dt.a.l++;
+		sll_allocator_resize((void**)(&(o->dt.a.v)),o->dt.a.l*sizeof(sll_object_t*));
+		o->dt.a.v[o->dt.a.l-1]=_build_single(&t,&tl,va);
+	} while (tl);
+	return o;
+}