+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Implemented `sll_get_cpu_index` and `sll_get_thread_index`
 - Implemented `SLL_IDENTIFIER_GET_STRING_INDEX`, `SLL_IDENTIFIER_IS_TLS`, `SLL_IDENTIFIER_UPDATE_STRING_INDEX`, `SLL_IDENTIFIER_SET_STRING_INDEX`, `sll_identifier_get_string_index`, `sll_identifier_is_tls`, `sll_identifier_set_string_index` and `sll_identifier_update_string_index` ([#254])
+- Implemented `sll_new_object` and `sll_new_object_list` ([#255])
 - Implemented proper parsing of unknown escape sequences
 - Operator parser is now generated from [`src/sll/data/operator_parser.txt`][0.7.10/src/sll/data/operator_parser.txt]
 - String format that converts objects to string: `S`
@@ -1821,6 +1822,7 @@ Unfortunately, no versions were assigned before 2021-06-15 (:disappointed:), so 
 [0.1.4]: https://github.com/sl-lang/sll/compare/lll-v0.1.3...lll-v0.1.4
 [0.1.3]: https://github.com/sl-lang/sll/compare/lll-v0.1.2...lll-v0.1.3
 [0.1.2]: https://github.com/sl-lang/sll/tree/v0.1.2
+[#255]: https://github.com/sl-lang/sll/issues/255
 [#254]: https://github.com/sl-lang/sll/issues/254
 [#247]: https://github.com/sl-lang/sll/issues/247
 [#246]: https://github.com/sl-lang/sll/issues/246